// packages/api/src/api/types/generated.ts
export type Access = AccessEntry[];

/**
 * AccessEntry represents an entity having access to the resource.
 *
 *
 * API extension: instance_access.
 */
export interface AccessEntry {
  /**
   * Certificate fingerprint
   * @example 636b69519d27ae3b0e398cb7928043846ce1e3842f0ca7a589993dd913ab8cc9
   */
  identifier: string;
  /**
   * The role associated with the certificate
   * @example admin, view, operator
   */
  role: string;
  /**
   * Which authorization method the certificate uses
   * @example tls, openfga
   */
  provider: string;
}

/**
 * BackupTarget represents the target storage server for an instance or volume backup.
 *
 *
 * API extension: backup_s3_upload.
 */
export interface BackupTarget {
  /**
   * Protocol is the upload protocol.
   * @example S3
   */
  protocol: string;
  /**
   * URL is the HTTPS URL for the backup
   * @example https://storage.googleapis.com
   */
  url: string;
  /**
   * BucketName is the name of the S3 bucket.
   * @example my_bucket
   */
  bucket_name: string;
  /**
   * Path is the target path.
   * @example foo/test.tar
   */
  path: string;
  /**
   * AccessKey is the S3 API access key
   * @example GOOG1234
   */
  access_key: string;
  /**
   * SecretKey is the S3 API access key
   * @example secret123
   */
  secret_key: string;
}
// ...
export type ConfigMap = Record<string, string>;

/**
 * DevicesMap type is used to hold incus devices configurations. In contrast to
 * plain map[string]map[string]string it provides unmarshal methods for JSON and
 * YAML, which gracefully handle numbers and bools.
 *
 *
 * @example {"eth0": {"network": "incusbr0", "type": "nic"}}
 */
export type DevicesMap = Record<string, Record<string, string>>;
// ...
export interface Instance extends InstancePut {
  /**
   * Instance creation timestamp
   * @example 2021-03-23T20:00:00-04:00
   */
  created_at: string;
  /**
   * Expanded configuration (all profiles and local config merged)
   * @example {"security.nesting": "true"}
   */
  expanded_config?: ConfigMap;
  /**
   * Expanded devices (all profiles and local devices merged)
   * @example {"root": {"type": "disk", "pool": "default", "path": "/"}}
   */
  expanded_devices?: DevicesMap;
  /**
   * Instance name
   * @example foo
   */
  name: string;
  /**
   * Instance status (see instance_state)
   * @example Running
   */
  status: string;
  /**
   * Instance status code (see instance_state)
   * @example 101
   */
  status_code: StatusCode;
  /**
   * Last start timestamp
   * @example 2021-03-23T20:00:00-04:00
   */
  last_used_at: string;
  /**
   * What cluster member this instance is located on
   * @example server01
   */
  location: string;
  /**
   * The type of instance (container or virtual-machine)
   * @example container
   */
  type: string;
  /**
   * Instance project name
   * @example foo
   *
   * API extension: instance_all_projects
   */
  project: string;
}

/**
 * InstanceBackup represents an instance backup.
 *
 *
 * API extension: instances.
 */
export interface InstanceBackup {
  /**
   * Backup name
   * @example backup0
   */
  name: string;
  /**
   * When the backup was created
   * @example 2021-03-23T16:38:37.753398689-04:00
   */
  created_at: string;
  /**
   * When the backup expires (gets auto-deleted)
   * @example 2021-03-23T17:38:37.753398689-04:00
   */
  expires_at: string;
  /**
   * Whether to ignore snapshots
   * @example false
   */
  instance_only: boolean;
  /**
   * Whether to use a pool-optimized binary format (instead of plain tarball)
   * @example true
   */
  optimized_storage: boolean;
}

/**
 * InstanceBackupPost represents the fields available for the renaming of a instance backup.
 *
 *
 * API extension: instances.
 */
export interface InstanceBackupPost {
  /**
   * New backup name
   * @example backup1
   */
  name: string;
}

/**
 * InstanceBackupsPost represents the fields available for a new instance backup.
 *
 *
 * API extension: instances.
 */
export interface InstanceBackupsPost {
  /**
   * Backup name
   * @example backup0
   */
  name: string;
  /**
   * When the backup expires (gets auto-deleted)
   * @example 2021-03-23T17:38:37.753398689-04:00
   */
  expires_at: string;
  /**
   * Whether to ignore snapshots
   * @example false
   */
  instance_only: boolean;
  /**
   * Whether to ignore dependent volumes
   * @example false
   */
  root_only: boolean;
  /**
   * Whether to use a pool-optimized binary format (instead of plain tarball)
   * @example true
   */
  optimized_storage: boolean;
  /**
   * What compression algorithm to use
   * @example gzip
   *
   * API extension: backup_compression_algorithm
   */
  compression_algorithm: string;
  /**
   * External upload target
   * The backup will be uploaded and then deleted from local storage.
   *
   * API extension: backup_s3_upload
   */
  target: BackupTarget | null;
}
// ...
export interface InstanceFull extends Instance {
  /**
   * List of backups.
   */
  backups: InstanceBackup[];
  /**
   * Current state.
   */
  state: InstanceState | null;
  /**
   * List of snapshots.
   */
  snapshots: InstanceSnapshot[];
}
// ...
export interface InstancePost {
  /**
   * New name for the instance
   * @example bar
   */
  name: string;
  /**
   * Whether the instance is being migrated to another server
   * @example false
   */
  migration: boolean;
  /**
   * Whether to perform a live migration (migration only)
   * @example false
   */
  live: boolean;
  /**
   * Whether to transfer a running instance incrementally
   * @example false
   *
   * API extension: instance_refresh_migration
   */
  refresh: boolean;
  /**
   * Whether snapshots should be discarded (migration only)
   * @example false
   */
  instance_only: boolean;
  /**
   * Target for the migration, will use pull mode if not set (migration only)
   */
  target: InstancePostTarget | null;
  /**
   * Target pool for local cross-pool move
   * @example baz
   *
   * API extension: instance_pool_move
   */
  pool: string;
  /**
   * Target project for local cross-project move
   * @example foo
   *
   * API extension: instance_project_move
   */
  project: string;
  /**
   * AllowInconsistent allow inconsistent copies when migrating.
   * @example false
   *
   * API extension: instance_allow_inconsistent_copy
   */
  allow_inconsistent: boolean;
  /**
   * Instance configuration file.
   * @example {"security.nesting": "true"}
   *
   * API extension: instance_move_config
   */
  Config: ConfigMap;
  /**
   * Instance devices.
   * @example {"root": {"type": "disk", "pool": "default", "path": "/"}}
   *
   * API extension: instance_move_config
   */
  Devices: DevicesMap;
  /**
   * List of profiles applied to the instance.
   * @example ["default"]
   *
   * API extension: instance_move_config
   */
  Profiles: string[];
}

/**
 * InstancePostTarget represents the migration target host and operation.
 *
 *
 * API extension: instances.
 */
export interface InstancePostTarget {
  /**
   * The certificate of the migration target
   * @example X509 PEM certificate
   */
  certificate: string;
  /**
   * The operation URL on the remote target
   * @example https://1.2.3.4:8443/1.0/operations/5e8e1638-5345-4c2d-bac9-2c79c8577292
   */
  operation?: string;
  /**
   * Migration websockets credentials
   * @example {"migration": "random-string", "criu": "random-string"}
   */
  secrets?: Record<string, string>;
}

/**
 * InstancePut represents the modifiable fields of an instance.
 *
 *
 * API extension: instances.
 */
export interface InstancePut {
  /**
   * Architecture name
   * @example x86_64
   */
  architecture: string;
  /**
   * Instance configuration (see doc/instances.md)
   * @example {"security.nesting": "true"}
   */
  config: ConfigMap;
  /**
   * Instance devices (see doc/instances.md)
   * @example {"root": {"type": "disk", "pool": "default", "path": "/"}}
   */
  devices: DevicesMap;
  /**
   * Whether the instance is ephemeral (deleted on shutdown)
   * @example false
   */
  ephemeral: boolean;
  /**
   * List of profiles applied to the instance
   * @example ["default"]
   */
  profiles: string[];
  /**
   * If set, instance will be restored to the provided snapshot name
   * @example snap0
   */
  restore?: string;
  /**
   * Whether only the instances disk should be restored
   * @example false
   */
  disk_only?: boolean;
  /**
   * Whether the instance currently has saved state on disk
   * @example false
   */
  stateful: boolean;
  /**
   * Instance description
   * @example My test instance
   */
  description: string;
}

/**
 * InstanceRebuildPost indicates how to rebuild an instance.
 *
 *
 * API extension: instances_rebuild.
 */
export interface InstanceRebuildPost {
  /**
   * Rebuild source
   */
  source: InstanceSource;
}

/**
 * InstanceSnapshot represents an instance snapshot.
 *
 *
 * API extension: instances.
 */
export interface InstanceSnapshot extends InstanceSnapshotPut {
  /**
   * Architecture name
   * @example x86_64
   */
  architecture: string;
  /**
   * Instance configuration (see doc/instances.md)
   * @example {"security.nesting": "true"}
   */
  config: ConfigMap;
  /**
   * Instance creation timestamp
   * @example 2021-03-23T20:00:00-04:00
   */
  created_at: string;
  /**
   * Instance description
   * @example My description
   */
  description: string;
  /**
   * Instance devices (see doc/instances.md)
   * @example {"root": {"type": "disk", "pool": "default", "path": "/"}}
   */
  devices: DevicesMap;
  /**
   * Whether the instance is ephemeral (deleted on shutdown)
   * @example false
   */
  ephemeral: boolean;
  /**
   * Expanded configuration (all profiles and local config merged)
   * @example {"security.nesting": "true"}
   */
  expanded_config?: ConfigMap;
  /**
   * Expanded devices (all profiles and local devices merged)
   * @example {"root": {"type": "disk", "pool": "default", "path": "/"}}
   */
  expanded_devices?: DevicesMap;
  /**
   * Last start timestamp
   * @example 2021-03-23T20:00:00-04:00
   */
  last_used_at: string;
  /**
   * Snapshot name
   * @example foo
   */
  name: string;
  /**
   * List of profiles applied to the instance
   * @example ["default"]
   */
  profiles: string[];
  /**
   * Whether the instance currently has saved state on disk
   * @example false
   */
  stateful: boolean;
  /**
   * Size of the snapshot in bytes
   * @example 143360
   *
   * API extension: snapshot_disk_usage
   */
  size: number;
}
// ...
export interface InstanceSnapshotPut {
  /**
   * When the snapshot expires (gets auto-deleted)
   * @example 2021-03-23T17:38:37.753398689-04:00
   */
  expires_at: string;
}
// ...
export interface InstanceState {
  /**
   * Current status (Running, Stopped, Frozen or Error)
   * @example Running
   */
  status: string;
  /**
   * Numeric status code (101, 102, 110, 112)
   * @example 101
   */
  status_code: StatusCode;
  /**
   * Disk usage key/value pairs
   */
  disk: Record<string, InstanceStateDisk>;
  /**
   * Memory usage information
   */
  memory: InstanceStateMemory;
  /**
   * Network usage key/value pairs
   */
  network: Record<string, InstanceStateNetwork>;
  /**
   * PID of the runtime
   * @example 7281
   */
  pid: number;
  /**
   * Number of processes in the instance
   * @example 50
   */
  processes: number;
  /**
   * CPU usage information
   */
  cpu: InstanceStateCPU;
  /**
   * The time that the instance started at
   *
   * API extension: instance_state_started_at.
   */
  started_at: string;
  /**
   * OS information.
   *
   * API extension: instances_state_os_info.
   */
  os_info: InstanceStateOSInfo | null;
}

/**
 * InstanceStateCPU represents the cpu information section of an instance's state.
 *
 *
 * API extension: instances.
 */
export interface InstanceStateCPU {
  /**
   * CPU usage in nanoseconds
   * @example 3637691016
   */
  usage: number;
  /**
   * CPU time available per second, in nanoseconds
   * @example 4000000000
   *
   * API extension: instance_state_cpu_time
   */
  allocated_time: number;
}

/**
 * InstanceStateDisk represents the disk information section of an instance's state.
 *
 *
 * API extension: instances.
 */
export interface InstanceStateDisk {
  /**
   * Disk usage in bytes
   * @example 502239232
   */
  usage: number;
  /**
   * Total size in bytes
   * @example 502239232
   *
   * API extension: instances_state_total
   */
  total: number;
}

/**
 * InstanceStateMemory represents the memory information section of an instance's state.
 *
 *
 * API extension: instances.
 */
export interface InstanceStateMemory {
  /**
   * Memory usage in bytes
   * @example 73248768
   */
  usage: number;
  /**
   * Peak memory usage in bytes
   * @example 73785344
   */
  usage_peak: number;
  /**
   * Total memory size in bytes
   * @example 12297557
   *
   * API extension: instances_state_total
   */
  total: number;
  /**
   * SWAP usage in bytes
   * @example 12297557
   */
  swap_usage: number;
  /**
   * Peak SWAP usage in bytes
   * @example 12297557
   */
  swap_usage_peak: number;
}
// ...
export interface InstanceStatePut {
  /**
   * State change action (start, stop, restart, freeze, unfreeze)
   * @example start
   */
  action: string;
  /**
   * How long to wait (in s) before giving up (when force isn't set)
   * @example 30
   */
  timeout: number;
  /**
   * Whether to force the action (for stop and restart)
   * @example false
   */
  force: boolean;
  /**
   * Whether to store the runtime state (for stop)
   * @example false
   */
  stateful: boolean;
}
// ...
export type StatusCode = 100 | 101 | 102 | 103 | 104 | 105 | 106 | 107 | 108 | 109 | 110 | 111 | 112 | 113 | 200 | 400 | 401 | (number & {});